+++ package.json
@@ -59,7 +59,6 @@
     "@dnd-kit/core": "^6.0.8",
     "@dnd-kit/modifiers": "^6.0.1",
     "@dnd-kit/sortable": "^7.0.2",
-    "@emotion/css": "11.11.2",
     "@emotion/react": "11.11.1",
     "@emotion/styled": "11.11.0",
     "@fontsource/inter": "^5.0.8",
@@ -69,7 +68,6 @@
     "@mui/icons-material": "5.14.7",
     "@mui/joy": "5.0.0-beta.5",
     "@mui/material": "5.14.7",
-    "@mui/system": "5.14.7",
     "@mui/x-date-pickers": "6.12.1",
     "@tauri-apps/api": "1.4.0",
     "@types/pulltorefreshjs": "0.1.5",
