+++ src/components/LocalizationDatePicker.tsx
@@ -1,3 +1,4 @@
+import { MaterialTheme } from "@/components/MaterialTheme";
 import useMediaQuery from "@/utils/useMediaQuery";
 import {
   Button,
@@ -200,61 +201,63 @@ export const LocalizationDatePicker = forwardRef<
   };
 
   return (
-    <LocalizationProvider
-      localeText={
-        language === "de"
-          ? deDE.components.MuiLocalizationProvider.defaultProps.localeText
-          : undefined
-      }
-      dateAdapter={AdapterDateFns}
-      adapterLocale={localeMap[language]}
-    >
-      <DatePicker
-        {...rest}
-        minDate={minDate}
-        maxDate={maxDate}
-        onChange={handleChange}
-        slots={{
-          field: JoyDateField,
-          // @ts-ignore
-          openPickerButton: JoyOpenPickerButton,
-          actionBar: JoyActionBar,
-        }}
-        ref={ref}
-        value={value}
-        selectedSections={selectedSections}
-        onSelectedSectionsChange={setSelectedSections}
-        slotProps={{
-          textField: (props) => ({
-            fullWidth: true,
-            inputProps: {
-              ...props.inputProps,
-              "data-testid": `${ariaLabel} textfield`,
-              "aria-label": ariaLabel,
-            },
-            onDoubleClick: handleDoubleClick,
-            onKeyDown: handleKeyDown,
-          }),
-          openPickerButton: {
+    <MaterialTheme>
+      <LocalizationProvider
+        localeText={
+          language === "de"
+            ? deDE.components.MuiLocalizationProvider.defaultProps.localeText
+            : undefined
+        }
+        dateAdapter={AdapterDateFns}
+        adapterLocale={localeMap[language]}
+      >
+        <DatePicker
+          {...rest}
+          minDate={minDate}
+          maxDate={maxDate}
+          onChange={handleChange}
+          slots={{
+            field: JoyDateField,
             // @ts-ignore
-            "data-testid": `${ariaLabel} pickerbutton`,
-          },
-          actionBar: {
-            actions: ["clear"],
-          },
-          popper: {
-            modifiers: [
-              {
-                name: "flip",
-                options: {
-                  altBoundary: false,
-                  fallbackPlacements: ["right", "left"],
-                },
+            openPickerButton: JoyOpenPickerButton,
+            actionBar: JoyActionBar,
+          }}
+          ref={ref}
+          value={value}
+          selectedSections={selectedSections}
+          onSelectedSectionsChange={setSelectedSections}
+          slotProps={{
+            textField: (props) => ({
+              fullWidth: true,
+              inputProps: {
+                ...props.inputProps,
+                "data-testid": `${ariaLabel} textfield`,
+                "aria-label": ariaLabel,
               },
-            ],
-          },
-        }}
-      />
-    </LocalizationProvider>
+              onDoubleClick: handleDoubleClick,
+              onKeyDown: handleKeyDown,
+            }),
+            openPickerButton: {
+              // @ts-ignore
+              "data-testid": `${ariaLabel} pickerbutton`,
+            },
+            actionBar: {
+              actions: ["clear"],
+            },
+            popper: {
+              modifiers: [
+                {
+                  name: "flip",
+                  options: {
+                    altBoundary: false,
+                    fallbackPlacements: ["right", "left"],
+                  },
+                },
+              ],
+            },
+          }}
+        />
+      </LocalizationProvider>
+    </MaterialTheme>
   );
 });
